Divided routes.py into modules for seperations of task
